requirements.txt
python-dotenv
requests
pyudev